fix: Fixing the hoek and lodash module vulnerabilities
The hoek module for Node is vulnerable to prototype pollution. An attacker could exploit this vulnerability to modify the Object prototype which causes additional properties or changes to exist on all objects. It is used by an old joi module version. The related dependencies are now updated and the tests have been adapted because the format of Joi related objects changed.